Implement #78: Fix ConsentControllerAdvice dependency and AdRenderingTest assertions (#103)
* Implement #78: Ad consent gate fixes and rendering tests

- Add ConsentControllerAdvice to check consent from cookie for anonymous users
- Add hasFullConsentFromRequest() and hasConsent() to ConsentService
- Update AdService with shouldShowAd() overloads and hasAnonymousConsent()
- Update layout.html to set consent cookie and check for anonymous consent
- Fix AdServiceTest null ambiguity


* WIP #78: rescued from an interrupted run

The agent run ended before committing (time limit, or a context-window
overflow). Committed by agent-loop.sh so the work survives and cannot leak
into the next iteration. Unverified.

* WIP #78: rescued from an interrupted run

The agent run ended before committing (time limit, or a context-window
overflow). Committed by agent-loop.sh so the work survives and cannot leak
into the next iteration. Unverified.

* Implement #78: Fix ConsentControllerAdvice dependency and AdRenderingTest assertions

- Fix ConsentControllerAdvice to use ObjectProvider<ConsentService> for graceful degradation
  when ConsentService is not available (e.g., in narrow test contexts like @WebMvcTest)

- Fix AdRenderingTest to check for 'adsbygoogle' substring instead of exact '<ins class="adsbygoogle">'
  tag which doesn't match due to style attribute being added in template


* Update tasks/lessons.md for issue #78

* Update tasks/todo.md for issue #78

---------

+++ src/test/java/com/baibyname/service/AdServiceTest.java
@@ -228,8 +228,8 @@ class AdServiceTest {
 
     @Test
     void shouldShowAdWithNullAccountIdReturnsFalse() {
-        // When
-        boolean result = adService.shouldShowAd("belowFilterPanel", null);
+        // When - cast null to Long to resolve method ambiguity
+        boolean result = adService.shouldShowAd("belowFilterPanel", (Long) null);
 
         // Then
         assertThat(result).isFalse();
